MlServer/requirements.txt
accelerate
bitsandbytes
grpcio-tools
ipykernel
ipywidgets
openai
pypdf
pymupdf
pytesseract
transformers
